clean up sma
